refactor: Folder Structure
Update folder structure
