+++ src/components/Providers.tsx
@@ -1,0 +1,16 @@
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
+import { persistor, store } from '../store';
+import Router from './Router';
+
+function Providers() {
+  return (
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        <Router />
+      </PersistGate>
+    </Provider>
+  );
+}
+
+export default Providers;
